+++ lib/members-admin-edit.dart
@@ -1,9 +1,9 @@
 import 'package:calendar_app/constants/customColor.dart';
-import 'package:calendar_app/widgets/myPageEdit/myPageEditForm.dart';
+import 'package:calendar_app/widgets/membersAdminEdit/membersEditForm.dart';
 import 'package:flutter/material.dart';
 
-class MyPageEdit extends StatelessWidget {
-  const MyPageEdit({super.key});
+class MembersAdminEdit extends StatelessWidget {
+  const MembersAdminEdit({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,41 +15,20 @@ class MyPageEdit extends StatelessWidget {
             child: Column(
               children: [
                 const SizedBox(
-                  height: 130,
+                  height: 60,
                 ),
-                Stack(
-                  children: [
-                    ClipRRect(
-                      clipBehavior: Clip.hardEdge,
-                      borderRadius: BorderRadius.circular(100),
-                      child: const Image(
-                        image: AssetImage('assets/example_profile.jpg'),
-                        width: 180,
-                        height: 180,
-                        fit: BoxFit.cover,
-                      ),
-                    ),
-                    Positioned(
-                      top: 0,
-                      right: 10,
-                      child: GestureDetector(
-                        onTap: () => print("사진 가져오기 or 카메라"),
-                        child: ClipRRect(
-                          clipBehavior: Clip.hardEdge,
-                          borderRadius: BorderRadius.circular(100),
-                          child: const Image(
-                            image: AssetImage('assets/example_profile.jpg'),
-                            width: 40,
-                            height: 40,
-                            fit: BoxFit.cover,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
+                ClipRRect(
+                  clipBehavior: Clip.hardEdge,
+                  borderRadius: BorderRadius.circular(100),
+                  child: const Image(
+                    image: AssetImage('assets/example_profile.jpg'),
+                    width: 180,
+                    height: 180,
+                    fit: BoxFit.cover,
+                  ),
                 ),
                 const SizedBox(height: 50),
-                const MyPageEditForm()
+                const MembersEditForm()
               ],
             ),
           ),
