--- lib/widgets/calendarWidget.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CalendarWidget extends StatelessWidget {
-  const CalendarWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
